pyproject.toml
[project]
name = "macless-haystack-owntracks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=43.0.3",
    "pbkdf2>=1.3",
    "pycryptodome>=3.21.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "srp>=1.0.21",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q"]
